LayoutTests/imported/w3c/web-platform-tests/svg/path/property/test_style_flush_on_dom_api_with_d_property-expected.txt


FAIL getTotalLength() with d property assert_equals: the total length expected 10 but got 0
FAIL getPointAtLength() with d property assert_equals: x-axis position expected 10 but got 0
FAIL isPointInFill() with d property assert_equals: expected true but got false
FAIL isPointInStroke() with d property assert_equals: expected true but got false
